agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-genai
